G9-api/UserService/login/requirements.txt
Flask==2.2.3
flask_expects_json==1.7.0
jsonschema==4.17.3
psycopg2==2.9.5
psycopg2_binary==2.9.5
PyJWT==2.6.0
python-dotenv==1.0.0
python_bcrypt==0.3.2
waitress==2.1.2
